feat(search): same preview dropdown on the home search box
Pulls the panel out of the search page into SearchSuggest and drops it into the
hero's field too. The component owns the debounce, the fetch and the keyboard;
Enter with nothing highlighted and the 'All results' row submit the enclosing
form, so each caller keeps deciding what a full search means (run it in place,
or navigate to /search).

The hero's overflow-hidden moves onto a backdrop wrapper: the scaled blur still
needs clipping, the panel needs to hang past the bottom edge.
